change the sizes to usigned long int

+++ 0x00-hello_world/6-size.c
@@ -12,12 +12,12 @@ int main(void)
 	long long lli;
 	float f;
 
-	printf("Size of a char: %hu byte(s)\n", (unsigned short)sizeof(c));
-	printf("Size of an int: %hu byes(s)\n", (unsigned short)sizeof(i));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu byes(s)\n", (unsigned long)sizeof(i));
 	printf("Size of a long int: %lu byte(s)\n",
-	       (unsigned long int)sizeof(li));
-	printf("Size of a long long int: %llu byte(s)\n",
-	       (unsigned long long int)sizeof(lli));
-	printf("Size of a float: %ld byte(s)\n", (unsigned long int)sizeof(f));
+	       (unsigned long)sizeof(li));
+	printf("Size of a long long int: %lu byte(s)\n",
+	       (unsigned long)sizeof(lli));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
